Refactor BackendCommandLine class
